use servehttp in tests
