+++ internal/filequeue/file_queue_test.go
@@ -1,4 +1,4 @@
-package queue
+package filequeue
 
 import "io"
